+++ mpc_energy/ha_mqtt.py
@@ -39,9 +39,21 @@ class CreateSelectInput():
     def callback_function(self, client: Client, user_data, message: MQTTMessage):
         self.state = message.payload.decode()
         self.entity.select_option(self.state)
+    
+    def publish_command(self, command):
+        command_topic = getattr(self.entity, "_command_topic", None)
+        mqtt_client = getattr(self.entity, "mqtt_client", None)
+
+        if command_topic is not None and mqtt_client is not None:
+            mqtt_client.publish(command_topic, payload=command, qos=0, retain=True)
+        else:
+            logger.warning(f"Unable to publish command for {self.name}: missing MQTT command topic/client")
+
         
-    def set_state(self, state):
+    def set_state(self, state, publish_command=False):
         if(state in self.options):
+            if publish_command:
+                self.publish_command(state)
             self.entity.select_option(state)
             self.state = state
         else:
@@ -119,14 +131,7 @@ automatic_control_switch = CreateSwitchInput(
 energy_controller_selector = CreateSelectInput(
     name="Energy Controller",
     unique_id="energy_controller",
-    options=["MPC", "RBC", "Safe Mode"]
-)
-
-
-base_load_sensor = CreateSensor(
-    name = "Base Load",
-    unique_id="base_load_python",
-    unit_of_measurement="w"
+    options=["MPC", "Safe Mode"]
 )
 
 alive_time_sensor = CreateSensor(
@@ -207,13 +212,6 @@ profit_tomorrow_sensor = CreateSensor(
     unit_of_measurement="$"
 )
 
-
-target_discharge_sensor = CreateSensor(
-    name = "Target Discharge Price",
-    unique_id="target_discharge_price_python",
-    unit_of_measurement="c/kWh"
-)
-
 kwh_discharged_sensor = CreateSensor(
     name = "kWh Discharged",
     unique_id="kwh_discharged_python",
@@ -236,6 +234,12 @@ kwh_required_overnight_sensor = CreateSensor(
 kwh_required_till_sundown_sensor = CreateSensor(
     name = "kWh Till Sundown",
     unique_id="kwh_required_till_sundown_python",
+    unit_of_measurement="kWh"
+)
+
+next_grid_interaction_kwh_sensor = CreateSensor(
+    name = "Next Grid Interaction",
+    unique_id="next_grid_interaction_kwh_python",
     unit_of_measurement="kWh"
 )
 
@@ -284,6 +288,56 @@ control_mode_override_duration_selector = CreateSelectInput(
     options=["5", "15", "30", "60", "120", "240", "360", "Till Price Change"]
 )
 
+ev_charging_mode_selector = CreateSelectInput(
+    name="EV Charging Mode",
+    unique_id="ev_charging_mode",
+    options=[
+        "Charging Disabled",
+        "Solar Smart",
+        "Ready by Time",
+        "Force On",
+    ]
+)
+
+
+ready_by_time_selector = CreateSelectInput(
+    name="EV Ready By Time",
+    unique_id="ev_ready_by_time",
+    options=[
+        "NA",
+        "00:00", "00:30",
+        "01:00", "01:30",
+        "02:00", "02:30",
+        "03:00", "03:30",
+        "04:00", "04:30",
+        "05:00", "05:30",
+        "06:00", "06:30",
+        "07:00", "07:30",
+        "08:00", "08:30",
+        "09:00", "09:30",
+        "10:00", "10:30",
+        "11:00", "11:30",
+        "12:00", "12:30",
+        "13:00", "13:30",
+        "14:00", "14:30",
+        "15:00", "15:30",
+        "16:00", "16:30",
+        "17:00", "17:30",
+        "18:00", "18:30",
+        "19:00", "19:30",
+        "20:00", "20:30",
+        "21:00", "21:30",
+        "22:00", "22:30",
+        "23:00", "23:30"
+    ]
+)
+
+target_ev_charge_rate_sensor = CreateSensor(
+    name = "Target EV Charge Rate",
+    unique_id="target_ev_charge_rate",
+    unit_of_measurement="kW"
+)
+
 def initalise_entities(): # Initalise entities and get them discovered by the ha mqtt service
     automatic_control_switch.set_state(False)
     energy_controller_selector.set_state("MPC")
@@ -295,11 +349,10 @@ def initalise_entities(): # Initalise entities and get them discovered by the ha
     current_feedIn_sensor.set_state(0)
     current_general_price_sensor.set_state(0)
     max_feedIn_sensor.set_state(0)
-    target_discharge_sensor.set_state(0)
     kwh_discharged_sensor.set_state(0)
     kwh_remaining_sensor.set_state(0)
     effective_price_sensor.set_state(0)
-    base_load_sensor.set_state(0)
+    next_grid_interaction_kwh_sensor.set_state(0)
     avg_daily_load_sensor.set_state(0)
     kwh_required_till_sundown_sensor.set_state(0)
     estimated_price_status_sensor.set_state(0)
@@ -312,6 +365,9 @@ def initalise_entities(): # Initalise entities and get them discovered by the ha
     control_mode_override_duration_selector.set_state("15")
     curtailment_status_sensor.set_state(0)
     curtailment_reason_sensor.set_state("None")
+    target_ev_charge_rate_sensor.set_state(0)
+    ev_charging_mode_selector.set_state("Solar Smart")
+    ready_by_time_selector.set_state("NA")
 
     time.sleep(10)