+++ mpc_energy/main.py
@@ -79,12 +79,13 @@ def start_streamlit_dashboard():
         "--theme.base=light"
     ])
 
-def send_mobile_notification(title, message):
+def send_mobile_notification(title, message, channel=None):
     try:
         ha.send_notification(
             title=title,
             message=message,
-            target=config_manager.notification_target
+            target=config_manager.notification_target,
+            channel=channel
         )
     except Exception as notification_error:
         logger.error(f"Failed to send mobile notification. This likely means that the notification target is incorrect. Check the notification target and try again. Error sending notification: {notification_error}")
@@ -103,7 +104,7 @@ def PrintError(e):
         )
         if(config_manager.notification_target_option in ["error_warning", "both"] and config_manager.notification_target != "" and time.time() - last_error_mobile_notification_timestamp > 60*60): # If the user has selected to receive error warnings and it's been more than 60 minutes since the last error notification, send a new one
             last_error_mobile_notification_timestamp = time.time()
-            send_mobile_notification(title="MPC Energy Error", message=f"An error occurred: {e}. Check the MPC Energy Log for details.")
+            send_mobile_notification(title="MPC Energy Error", message=f"An error occurred: {e}. Check the MPC Energy Log for details.", channel="mpc_energy_error_warnings")
             
 
     except Exception as notification_error:
@@ -125,7 +126,6 @@ def FailSafe(e):
 
 while(started == False):
     try:
-        from RBC import RBC
         from MPC import MPC
         from energy_controller import EnergyController
         from ha_api import HomeAssistantAPI
@@ -175,21 +175,14 @@ while(started == False):
 
         control_mode_override_manager = ControlModeOverrideManager(ha_mqtt=ha_mqtt, energy_controller=EC, plant=plant)
 
-        rbc = RBC(
-            ha=ha, 
-            ha_mqtt=ha_mqtt,
-            plant=plant, 
-            EC=EC,
-            buffer_percentage_remaining=35, # percentage to inflate predicted load consumption
-        )
-
         mpc = MPC(
             ha=ha,
             plant=plant,
             EC=EC, 
             local_tz=ha.local_tz,
             demand_tarrif=demand_tariff,
-            retailer=config_manager.energy_retailer
+            retailer=config_manager.energy_retailer,
+            ha_mqtt=ha_mqtt
         ) 
 
         # Start Streamlit dashboard
@@ -221,19 +214,18 @@ def set_sensor_if_changed(sensor, value):
 
 # Update HA MQTT sensors
 def update_sensors(price_data):
-    rbc.update_values(amber_data=price_data)
-
     override_status = control_mode_override_manager.state['active']
     override_mode = control_mode_override_manager.state['mode']
     opperating_mode = override_mode if override_status else EC.working_mode
+
     set_sensor_if_changed(ha_mqtt.max_feedIn_sensor, round(price_data.feedIn_max_forecast_price))
     set_sensor_if_changed(ha_mqtt.current_feedIn_sensor, round(price_data.feedIn_price))
     set_sensor_if_changed(ha_mqtt.current_general_price_sensor, round(price_data.general_price))
     set_sensor_if_changed(ha_mqtt.kwh_discharged_sensor, round(plant.kwh_till_full, 2))
     set_sensor_if_changed(ha_mqtt.kwh_remaining_sensor, round(plant.kwh_stored_available, 2))
-    set_sensor_if_changed(ha_mqtt.target_discharge_sensor, round(rbc.target_dispatch_price))
-    set_sensor_if_changed(ha_mqtt.kwh_required_overnight_sensor, round(rbc.kwh_required_remaining, 2))    
-    set_sensor_if_changed(ha_mqtt.kwh_required_till_sundown_sensor, round(rbc.kwh_required_till_sundown, 2))
+    set_sensor_if_changed(ha_mqtt.kwh_required_overnight_sensor, round(plant.kwh_required_remaining(buffer_percentage=0), 2))    
+    set_sensor_if_changed(ha_mqtt.kwh_required_till_sundown_sensor, round(plant.kwh_required_till_sundown(buffer_percentage=0), 2))
+    set_sensor_if_changed(ha_mqtt.next_grid_interaction_kwh_sensor, round(mpc.next_grid_interaction_kwh, 2))
     set_sensor_if_changed(ha_mqtt.working_mode_sensor, opperating_mode)
 
     set_sensor_if_changed(ha_mqtt.import_cost_sensor, round(plant.daily_import_cost, 2))
@@ -241,6 +233,8 @@ def update_sensors(price_data):
     set_sensor_if_changed(ha_mqtt.net_profit_sensor, round(plant.daily_net_profit, 2))
     set_sensor_if_changed(ha_mqtt.profit_remaining_today_sensor, round(mpc.profit_remaining_today, 2))
     set_sensor_if_changed(ha_mqtt.profit_tomorrow_sensor, round(mpc.profit_tomorrow, 2))
+    if(mpc.ev_configured):
+       set_sensor_if_changed(ha_mqtt.target_ev_charge_rate_sensor, round(mpc.target_ev_charge_rate, 2))
 
     if(plant.grid_power < 0):
         price = price_data.feedIn_price
@@ -250,7 +244,6 @@ def update_sensors(price_data):
         grid_status = "I"
     
     set_sensor_if_changed(ha_mqtt.system_state_sensor, opperating_mode + f" {round(abs(plant.grid_power),1)}{grid_status}@{price} c/kWh ${round(plant.daily_net_profit,2)} profit")
-    set_sensor_if_changed(ha_mqtt.base_load_sensor, round(1000*plant.get_base_load_estimate(),2)) # converted to w from kW
     set_sensor_if_changed(ha_mqtt.effective_price_sensor, round(mpc.current_effective_price*100)) 
     set_sensor_if_changed(ha_mqtt.avg_daily_load_sensor, round(plant.avg_daily_load,2))
 
@@ -291,14 +284,16 @@ def check_for_spike(price_data):
                     notification_id="mpc_energy_price_spike_warning"
                 )
                 if(config_manager.notification_target_option in ["price_spike_warning", "both"] and config_manager.notification_target != ""):
-                    send_mobile_notification(title="MPC Forecast Feed In Price Spike", message=spike_message)   
+                    send_mobile_notification(title="MPC Forecast Feed In Price Spike", message=spike_message, channel="mpc_energy_price_spike_warnings")   
             
         else:
             spike_found_timestamp = 0 # Reset the spike found timestamp if no spikes are currently forecasted
                 
+last_ev_charging_mode = ha_mqtt.ev_charging_mode_selector.state
+last_ev_charged_by_time = ha_mqtt.ready_by_time_selector.state                
 
 def run_controller(price_update=False):
-    global automatic_control, last_control_mode
+    global automatic_control, last_control_mode, last_ev_charging_mode, last_ev_charged_by_time
     # If Auto control has been TURNED on, print a msg and reset flag
     selected_controller = ha_mqtt.energy_controller_selector.state
 
@@ -308,14 +303,22 @@ def run_controller(price_update=False):
             mpc.run_optimisation(price_data) # Run the MPC optimisation each time the price updates to keep the plot updated
         return
     
+    if(ha_mqtt.ev_charging_mode_selector.state != last_ev_charging_mode or ha_mqtt.ready_by_time_selector.state != last_ev_charged_by_time):
+        last_ev_charging_mode = ha_mqtt.ev_charging_mode_selector.state
+        last_ev_charged_by_time = ha_mqtt.ready_by_time_selector.state
+           
+        logger.debug(f"EV Charging settings changed. Forcing MPC to update with new settings.")
+        mpc.run_optimisation(price_data) # Run the MPC optimisation to update the EV charging plan with the new settings
+    
+    if(ha_mqtt.ev_charging_mode_selector.state != "Ready by Time" and ha_mqtt.ready_by_time_selector.state != "None"):
+        ha_mqtt.ready_by_time_selector.set_state("NA") # Reset the ready by time selector if the user changes the mode to something other than ready by time to prevent confusion
+    
     if(last_control_mode == "Manual Override"):
         if(ha_mqtt.automatic_control_switch.state == True):
             logger.warning(f"Manual override finished. Returning control to {selected_controller}.")
 
             if(selected_controller == "MPC"):
                 mpc.run(price_data)
-            elif(selected_controller == "RBC"):
-                rbc.run(price_data)
             else:
                 EC.self_consumption()
 
@@ -335,9 +338,6 @@ def run_controller(price_update=False):
         if(selected_controller == "MPC"):
             if(last_control_mode != selected_controller or price_update == True):
                 mpc.run(price_data) # Run the MPC Controller if the price updates (every 5 min) or if it was just selected as the new controller
-
-        elif(selected_controller == "RBC"):
-            rbc.run(price_data) # RBC needs to run every loop
 
         else: # Selected Controller must be safe mode
             EC.self_consumption()